# src/main/resources/static/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: api-acervo-digital
  version: 1.0.0
  summary: Contrato OpenAPI do microsserviço api-acervo-digital.
  description: API responsavel pelo acervo digital da Plataforma Via Fluvial, incluindo upload, gestao e consulta de documentos e midias.
  contact:
    name: Via Fluvial
    url: https://viafluvial.com.br
servers:
  - url: /api/v1
    description: Runtime de Desenvolvimento
  - url: https://api-acervo-digital.hml.viafluvial.com.br/api/v1
    description: Runtime de Homologação
  - url: https://api-acervo-digital.prd.viafluvial.com.br/api/v1
    description: Runtime de Produção
security:
  - bearerAuth: []
...
paths:
  /acervo/media/uploads:
    post:
      tags:
        - Media
      operationId: uploadMedia
      summary: Enviar imagem ou vídeo para o acervo.
      description: |
        Recebe uma imagem ou vídeo e registra o item no Acervo Digital.
        O arquivo enviado entra inicialmente em quarentena e não deve ser publicado
        automaticamente. A mídia somente poderá ser exibida publicamente após aprovação
        administrativa e publicação.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MediaUploadRequest'
      responses:
        '201':
          description: Mídia recebida e registrada em quarentena.
          headers:
            Location:
              description: URL de consulta do recurso criado.
              schema:
                type: string
                example: /api/v1/acervo/media/med_01JABCDEF1234567890
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItemResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /acervo/media:
    get:
      tags:
        - Media
      operationId: searchMedia
      summary: Consultar mídias do acervo.
      description: |
        Retorna imagens e vídeos cadastrados no Acervo Digital com filtros de negócio.
        Pode ser usado pelo painel administrativo, por serviços internos e por fluxos
        que precisam listar mídias conforme entidade, status, finalidade e visibilidade.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/EntityTypeQuery'
        - $ref: '#/components/parameters/EntityIdQuery'
        - $ref: '#/components/parameters/MediaTypeQuery'
        - $ref: '#/components/parameters/PurposeQuery'
        - $ref: '#/components/parameters/VisibilityQuery'
        - $ref: '#/components/parameters/StatusQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
        - $ref: '#/components/parameters/SortQuery'
      responses:
        '200':
          description: Lista paginada de mídias.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedMediaResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /acervo/entities/{entityType}/{entityId}/media:
    get:
      tags:
        - Media
      operationId: listMediaByEntity
      summary: Listar mídias vinculadas a uma entidade.
      description: |
        Lista mídias associadas a uma entidade de negócio, como embarcação,
        acomodação, barqueiro, campanha ou página do site. Por padrão, os consumidores
        externos devem utilizar apenas mídias publicadas.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/EntityTypePath'
        - $ref: '#/components/parameters/EntityIdPath'
        - $ref: '#/components/parameters/OnlyPublishedQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: Mídias vinculadas à entidade.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedMediaResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /acervo/media/{mediaPublicKey}/approve:
    post:
      tags:
        - MediaApproval
      operationId: approveMedia
      summary: Aprovar mídia.
      description: |
        Aprova uma imagem ou vídeo após análise administrativa. A aprovação libera a
        mídia para a próxima etapa de publicação, mas não implica obrigatoriamente que
        ela já esteja disponível na CDN.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/MediaPublicKeyPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionRequest'
      responses:
        '200':
          description: Mídia aprovada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /acervo/media/{mediaPublicKey}/reject:
    post:
      tags:
        - MediaApproval
      operationId: rejectMedia
      summary: Reprovar mídia.
      description: |
        Reprova uma mídia em análise. A reprovação exige motivo padronizado e pode
        conter observação administrativa. Mídias reprovadas não podem ser exibidas
        publicamente nem utilizadas em marketing.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/MediaPublicKeyPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectionDecisionRequest'
      responses:
        '200':
          description: Mídia reprovada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /acervo/media/{mediaPublicKey}/publish:
    post:
      tags:
        - MediaApproval
      operationId: publishMedia
      summary: Publicar mídia aprovada.
      description: |
        Publica uma mídia previamente aprovada. Após a publicação, a mídia pode ser
        disponibilizada para consumo público conforme sua visibilidade, finalidade e
        URLs de CDN geradas.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/MediaPublicKeyPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishMediaRequest'
      responses:
        '200':
          description: Mídia publicada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /acervo/media/{mediaPublicKey}/block:
    post:
      tags:
        - MediaApproval
      operationId: blockMedia
      summary: Bloquear mídia publicada ou aprovada.
      description: |
        Bloqueia uma mídia por risco, reclamação, violação de direito de uso,
        inconsistência documental ou decisão administrativa. Mídias bloqueadas não
        devem aparecer no fluxo de venda, site, marketing ou integrações públicas.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/MediaPublicKeyPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockMediaRequest'
      responses:
        '200':
          description: Mídia bloqueada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /acervo/media/{mediaPublicKey}/archive:
    post:
      tags:
        - MediaApproval
      operationId: archiveMedia
      summary: Arquivar mídia.
      description: |
        Arquiva uma mídia que não deve mais ser usada operacionalmente. O arquivamento
        não representa exclusão física imediata e deve respeitar política de retenção.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/MediaPublicKeyPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArchiveRequest'
      responses:
        '200':
          description: Mídia arquivada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /acervo/media/{mediaPublicKey}/main:
    patch:
      tags:
        - Media
      operationId: setMainMedia
      summary: Definir mídia como principal.
      description: |
        Define ou remove a mídia como principal para a entidade vinculada. Quando uma
        nova mídia é definida como principal, as demais mídias da mesma entidade e
        finalidade podem perder a marcação de principal conforme regra de negócio.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/MediaPublicKeyPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetMainMediaRequest'
      responses:
        '200':
          description: Mídia principal atualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /acervo/media/{mediaPublicKey}/order:
    patch:
      tags:
        - Media
      operationId: updateMediaOrder
      summary: Atualizar ordem de exibição da mídia.
      description: |
        Atualiza a posição de exibição da mídia dentro da galeria ou lista da entidade
        relacionada.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/MediaPublicKeyPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
      responses:
        '200':
          description: Ordem atualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /acervo/private-documents/uploads:
    post:
      tags:
        - Documents
      operationId: uploadPrivateDocument
      summary: Enviar documento privado.
      description: |
        Recebe documento privado de barqueiro, embarcação, agência ou outro contexto
        administrativo. Documentos privados nunca devem ser expostos em CDN pública.
        O acesso deve ser autenticado ou realizado por URL assinada temporária.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PrivateDocumentUploadRequest'
      responses:
        '201':
          description: Documento recebido e registrado.
          headers:
            Location:
              description: URL de consulta do documento criado.
              schema:
                type: string
                example: /api/v1/acervo/private-documents/doc_01JABCDEF1234567890
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateDocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /acervo/private-documents:
    get:
      tags:
        - Documents
      operationId: searchPrivateDocuments
      summary: Consultar documentos privados.
      description: |
        Retorna documentos privados conforme filtros de entidade, tipo documental,
        status e validade. O retorno não deve expor URL pública permanente do arquivo.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/EntityTypeQuery'
        - $ref: '#/components/parameters/EntityIdQuery'
        - name: documentTypeCode
          in: query
          required: false
          description: Código do tipo documental.
          schema:
            type: string
            example: LICENCA_OPERACAO
        - name: status
          in: query
          required: false
          description: Status do documento.
          schema:
            $ref: '#/components/schemas/DocumentStatus'
        - name: expiresUntil
          in: query
          required: false
          description: Filtra documentos com validade até a data informada.
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: Lista paginada de documentos privados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedPrivateDocumentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /acervo/private-documents/{documentPublicKey}:
    get:
      tags:
        - Documents
      operationId: getPrivateDocumentByPublicKey
      summary: Consultar documento privado.
      description: |
        Retorna os metadados de um documento privado, incluindo vínculo de negócio,
        status, validade e histórico resumido. O conteúdo do arquivo não é retornado
        neste endpoint.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/DocumentPublicKeyPath'
      responses:
        '200':
          description: Documento localizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateDocumentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Documents
      operationId: removePrivateDocument
      summary: Remover documento privado.
      description: |
        Remove um documento privado do acervo: marca o documento como REMOVIDO e
        elimina o arquivo físico do bucket privado quando existir.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/DocumentPublicKeyPath'
      responses:
        '200':
          description: Documento removido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /acervo/private-documents/{documentPublicKey}/approve:
    post:
      tags:
        - DocumentApproval
      operationId: approvePrivateDocument
      summary: Aprovar documento privado.
      description: |
        Aprova um documento privado após análise administrativa ou de compliance.
        Documentos aprovados podem liberar fluxos operacionais conforme regras de
        outros domínios, mas permanecem privados.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/DocumentPublicKeyPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionRequest'
      responses:
        '200':
          description: Documento aprovado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /acervo/document-types:
    get:
      tags:
        - Catalogs
      operationId: listDocumentTypes
      summary: Listar tipos documentais exigidos.
      description: |
        Retorna o catálogo de tipos de documentos aceitos pelo Acervo Digital,
        informando escopo, obrigatoriedade, necessidade de validade e exigência de
        aprovação.
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - name: scope
          in: query
          required: false
          description: Escopo do tipo documental.
          schema:
            $ref: '#/components/schemas/DocumentScope'
        - name: active
          in: query
          required: false
          description: Filtra tipos ativos ou inativos.
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: Tipos documentais retornados.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentTypeResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Identificador de correlação da requisição para rastreabilidade.
      schema:
        type: string
        maxLength: 100
        example: 4f5d6e7a-8b9c-40de-9f12-34567890abcd
...
  responses:
    BadRequest:
      description: Requisição inválida.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Token ausente, inválido ou expirado.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Usuário autenticado sem permissão para a operação.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    PayloadTooLarge:
      description: Arquivo enviado excede o tamanho permitido.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...